Working tables and plotting
The DataTable and plotting works in harmony now! A user can change the parameters in the table and it'll update the plot accordingly with the recalculated spectrum.

TODO - optimize performance; right now the spectrum updating is quite slow and I suspect it has to do with the synthetic spectrum generation (it has a for loop). Try and implement a list comp instead...